requirements.txt
Django==1.11.6
djangorestframework==3.7.0
requests==2.18.4